Borders and nav bar
Fixed the issue with borders not showing by removing the commas in the
CSS.
Fixed the issue with the nav bar not being on the side of the page.
Set margins on the header, nav bar, and article.
Created an article CSS class.
Added an image that might be used for the title bar or background (if
allowed).
Changed some wording for clarity.