Created folders, footer, and signup page
Seperated out images from code into two folders.
Created a footer and applied it to all pages.
Touched up the nav bar.
Created a tournament signup page that will email me with the person's
info. This page will be displayed instead of a bracket page until the
tournament begins. A PHP script is needed to actually make the form
email me the information.