Anti-spam and verify registration in PHP
This will be the last commit before going live. Any commits after this
should be made on a seperate branch.
Added an anti-spam input on the registration page.
Added some form validation in the PHP.
Made the unused mail function parameters null to prevent unwanted
forwarding.
Reformatted the registration page.
Added a tab icon that worked locally but not on the server.
Grammatical corrections.